requirements.txt
astroid==2.6.5
attrs==19.3.0
Automat==0.8.0
autopep8==1.5.7
blessed==1.17.6
blinker==1.4
botocore==1.20.87
breezy==3.0.2
brz-debian==3.0.0
certifi==2019.11.28
chalice==1.23.0
chardet==3.0.4
click==8.0.1
cloud-init==21.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
Deprecated==1.2.7
devscripts===2.20.2ubuntu2
distro==1.4.0
distro-info===0.23ubuntu1
dulwich==0.19.15
entrypoints==0.3
fastimport==0.9.8
Flask==2.0.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
gpg===1.13.1-unknown
greenlet==1.1.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.7
importlib-metadata==1.5.0
incremental==16.10.1
inquirer==2.7.0
isort==5.9.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
marshmallow==3.13.0
mccabe==0.6.1
more-itertools==4.2.0
mypy-extensions==0.4.3
mysqlclient==2.0.3
netifaces==0.10.4
numpy==1.21.2
oauthlib==3.1.0
opencv-python==4.5.3.56
pdf2image==1.16.0
pexpect==4.6.0
Pillow==8.3.1
py3dns==3.2.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.7.0
pycurl==7.43.0.2
PyGithub==1.43.7
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.9.5
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
PySimpleSOAP==1.16.2
pytesseract==0.3.8
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-debianbts==3.0.2
python-dotenv==0.19.0
python-editor==1.0.4
python-gitlab==2.0.1
python-magic==0.4.16
pyxdg==0.26
PyYAML==5.3.1
readchar==2.0.1
reportbug===7.6.0ubuntu1
requests==2.20.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.1
SQLAlchemy==1.4.23
ssh-import-id==5.10
systemd-python==234
termcolor==1.1.0
toml==0.10.2
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ubuntu-dev-tools==0.176
ufw==0.36
unattended-upgrades==0.1
unidiff==0.5.5
urllib3==1.24.3
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==1.0.0
zope.interface==4.7.1
